Moved non-subclass-specific tests to their own module, and added unit tests for Menu.create_dict_from_parent_context()
